Cairo::FtFontFace: Keep the FT_Face alive long enough to avoid segfaults
Link the lifetime of the FT_Face to that of the newly created cairo_font_face_t
by initially increasing the reference count of the SV wrapping the FT_Face and
decreasing it when the cairo_font_face_t is destroyed.

Test this with a unit test provided by Cornelius.

Also remove FT_Face typemap in favor of manual unwrapping in the only place
it's used.