Cairo::Surface: fix the exception conversion in callbacks
The recent fix of the off-by-one error in enum handling revealed a bug
in the exception conversion from strings to cairo_status_t in the
read/write callback marshallers.  When you do C<die 'no-memory'>, $@ is
actually filled with 'no-memory at foo.pl line 23\n'.  The off-by-one
error meant that this was converted successfully as if it was just
'no-memory'.  This is not the case anymore, and so we need to strip off
the location suffix before trying to convert to a cairo_status_t.